feat: Support MessageVenue

+++ app/src/main/java/com/mocharealm/compound/domain/model/Venue.kt
@@ -1,7 +1,7 @@
 package com.mocharealm.compound.domain.model
 
 data class Venue(
-    val longitude: Long,
-    val latitude: Long,
+    val longitude: Double,
+    val latitude: Double,
     val name: String,
 )
